requirements.txt
beautifulsoup4==4.9.1
requests==2.25.1
python-dateutil==2.8.1
